mmc: sdhci: Use quirks2 for defining new quirks
Revert back quirks to 32 bits long and use quirks2 to define any new
required quirks.

Bug 1330567

Reviewed-on: http://git-master/r/260564
(cherry picked from commit 836c16ce754c154856369cc2d5494ece77aa573c)

Reviewed-on: http://git-master/r/269189
GVS: Gerrit_Virtual_Submit

+++ drivers/mmc/host/sdhci-pltfm.h
@@ -17,7 +17,7 @@
 
 struct sdhci_pltfm_data {
 	const struct sdhci_ops *ops;
-	u64 quirks;
+	u32 quirks;
 	u32 quirks2;
 };
